move rlp util to be not private
